Perft

+++ src/Engine/General/Constants.cs
@@ -29,7 +29,7 @@ public static class Constants
 
     public const int ColourBitOffset = 3;
     public const int LastPlyMoveBitOffset = 6;
-    public const int PawnMoved2RanksFlag = 0b0010_0000;
-    public const int PieceKindMask = 0b0000_0111;
-    public const int PieceDescriptionMask = 0b0001_1111;
+    public const ushort PawnMoved2RanksFlag = 0b0010_0000;
+    public const ushort PieceKindMask = 0b0000_0111;
+    public const ushort PieceDescriptionMask = 0b0001_1111;
 }
